Added protection code if no handler was found.
Fixed a bug where the interpreter used the old values for stack pointer
and instruction pointer registers after an exception was triggered.

The Eco compiler is now also under an exception protection.
I noticed a conflict with the VM's GC mechanism if the exceptions
occurs within a nested function since the exception handling
mechanism modifies the pointer to the previous frame, making the GC
think that the discarded frames in between consist entirely of
Eco_Any values. This breaks the system!

Test case: { x = xxx }.